แก้ type double

+++ lib/pages/customer/CustomerProfile.dart
@@ -1023,7 +1023,7 @@ class _ProfilePageState extends State<ProfilePage> {
     if (cus_balance.statusCode == 200) {
       final data = jsonDecode(cus_balance.body);
       final int balance = data['balance'] ?? 0;
-      context.read<ShareData>().user_info_send.balance = balance;
+      context.read<ShareData>().user_info_send.balance = balance.toDouble();
     } else {
       Fluttertoast.showToast(msg: "โหลดยอดเงินไม่สำเร็จ");
     }
